+++ packages/node/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 - Use `ConnectionPoolStateManager` from host when workers are turned on
 
+### Added
+- `store-cache-upper-limit` flag to control limiting the max size of the store cache (#1859)
+
 ## [2.8.0] - 2023-06-26
 ### Added
 - import `PgMmrCacheService` into indexer module (#1828)